+++ CHANGELOG.md
@@ -9,8 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - LLM evaluation logic
+- Integrated Alembic for managing chat history migrations
 - Tool implementations without langchain or langgraph dependencies
 - Unit tests for the migrated tools
+- CRUDs.
+- BlueNaas CRUD tools
+
+### Fixed
+- Migrate LLM Evaluation logic to scripts and add tests
 
 ## [0.3.3] - 30.10.2024